requirements.txt
redis==5.0.1
flask==2.3.3
flask-cors==4.0.0
flask-httpauth==4.8.0
python-dotenv==1.0.0
uuid==1.30
click==8.1.7